chore: bump dalamud sdk to 15.0.0 and version to 0.5.0.0 (#10)

+++ AllaganSync/Tracking/Trackers/MonsterSpawnTracker.cs
@@ -99,7 +99,7 @@ public unsafe class MonsterSpawnTracker : IGameEventTracker
         spawnNpcHook!.Original(targetId, packet);
     }
 
-    private void OnTerritoryChanged(ushort territoryId)
+    private void OnTerritoryChanged(uint territoryId)
     {
         seenSpawns.Clear();
     }
